requirements.txt
aiogram==3.13.1
sqlalchemy==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0